Move share panel to another class so it does not crash the main window on Windows 7
