+++ AmiKoWindows/packages.config
@@ -7,6 +7,7 @@
   <package id="MahApps.Metro.Resources" version="0.6.1.0" targetFramework="net461" />
   <package id="MicroSoft.Net.Compilers" version="2.8.0" targetFramework="net461" />
   <package id="MMaitre.MediaCaptureWPF" version="1.0.0" targetFramework="net461" />
+  <package id="System.Runtime.WindowsRuntime" version="4.3.0" targetFramework="net461" />
   <package id="System.Data.SQLite" version="1.0.104.0" targetFramework="net461" />
   <package id="System.Data.SQLite.Core" version="1.0.104.0" targetFramework="net461" />
   <package id="System.Data.SQLite.EF6" version="1.0.104.0" targetFramework="net461" />
